requirements.txt
python-dotenv # For reading environment variables stored in .env file
langchain
langchain-community
langchain-core
langchain-ollama
langchain-openai # For embeddings
unstructured # Document loading
chromadb # Vector storage
openai # For embeddings
tiktoken  # For embeddings
requests
bs4
python-dotenv
nltk
gradio
flask
flask-cors
apscheduler
setuptools
PyYAML
routes
# install markdown depenendies with: `pip install "unstructured[md]"` after install the requirements file. Leave this line commented out.